Use folgezettel here
Because Traversable is conceptually a child of Foldable (given that it requires it in the constraint). This can be used to, eventually, model a tree of type classes.

The folgezettel syntax is backwards compatible with Obsidian, vscode-memo, etc. but is recognized only by neuron. It sill automatically render the tree on the generated site.